requirements.txt
Flask
Flask-SocketIO
mysql-connector-python
